[TFP-535] Move files into separate directories
