--- access_nix.go
@@ -1,30 +1,0 @@
-// +build !windows
-
-// Copyright 2018. Akamai Technologies, Inc
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import "golang.org/x/sys/unix"
-
-// Read/Write/Execute Constants
-const (
-	ACCESS_R_OK = unix.R_OK
-	ACCESS_W_OK = unix.W_OK
-	ACCESS_X_OK = unix.X_OK
-)
-
-func checkAccess(path string, perm uint32) error {
-	return unix.Access(path, perm)
-}
