+++ cli/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/akamai/cli/cli/app"
+	"os"
+)
+
+func main() {
+	os.Exit(app.Run())
+}
