--- upgrade_noop.go
@@ -1,33 +1,0 @@
-//+build noautoupgrade
-
-// Copyright 2018. Akamai Technologies, Inc
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-func checkUpgradeVersion(force bool) string {
-	return ""
-}
-
-func getLatestReleaseVersion() string {
-	return "0"
-}
-
-func upgradeCli(latestVersion string) bool {
-	return false
-}
-
-func getUpgradeCommand() *commandPackage {
-	return nil
-}
